requirements.txt
annotated-types==0.6.0
aspose-pdf==23.11.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numpy==1.24.4
packaging==23.2
preshed==3.0.9
pydantic==2.5.3
pydantic_core==2.14.6
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.2
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
